Added exception to delete image

+++ src/main/java/com/example/WeBKKHustraUhrovec/Service/impl/ImageIMPL.java
@@ -1,7 +1,9 @@
 package com.example.WeBKKHustraUhrovec.Service.impl;
 
 import com.example.WeBKKHustraUhrovec.Entity.Image;
+import com.example.WeBKKHustraUhrovec.Entity.News;
 import com.example.WeBKKHustraUhrovec.Repo.ImageRepo;
+import com.example.WeBKKHustraUhrovec.Repo.NewsRepo;
 import com.example.WeBKKHustraUhrovec.Service.ImageService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,28 +34,25 @@ public class ImageIMPL implements ImageService {
     private String fileUploadPath;
 
     @Autowired
-    private ResourceLoader resourceLoader;
+    private NewsRepo newsRepo;
 
     @Transactional
     @Override
     public Image addImage(MultipartFile file) {
         try {
             // Generate a unique filename
-            String fileName = UUID.randomUUID().toString() + "_" + file.getOriginalFilename();
+            String fileName = UUID.randomUUID() + "_" + file.getOriginalFilename();
 
-            // Create an ImageEntity
             Image imageEntity = new Image();
 
-            // Construct the file path using Paths.get() for better cross-platform compatibility
             Path filePath = Paths.get(fileUploadPath, fileName);
+
             imageEntity.setUrl(filePath.toString());
 
             imageEntity.setName(fileName);
 
-            // Save the ImageEntity to the database
             imageEntity = imageRepo.save(imageEntity);
 
-            // Save the file to the configured location
             Files.copy(file.getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
 
             return imageEntity;
@@ -75,19 +74,16 @@ public class ImageIMPL implements ImageService {
     @Override
     public byte[] getImage(String imageName) {
         try {
-            // Construct the full path of the image file
+
             String filePath = fileUploadPath + "/" + imageName;
 
-            // Log the file path for debugging
-            System.out.println("Image file path: " + filePath);
 
-            // Read the image bytes from the file
             Path imagePath = Paths.get(filePath);
             return Files.readAllBytes(imagePath);
         } catch (IOException e) {
-            // Log the exception for debugging
+
             e.printStackTrace();
-            // Handle exception (e.g., log it or throw a custom exception)
+
             throw new RuntimeException("Failed to load image", e);
         }
     }
@@ -95,18 +91,22 @@ public class ImageIMPL implements ImageService {
     @Transactional
     @Override
     public void deleteImage(Long id) {
-        // Find the image by ID
+
         Image image = imageRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("Image not found with id: " + id));
 
-        // Delete the image file
+        List<News> newsList = newsRepo.findAllByImage(image);
+        if (!newsList.isEmpty()) {
+            throw new RuntimeException("Cannot delete image because it is associated with news.");
+        }
+
         try {
             Files.delete(Paths.get(image.getUrl()));
         } catch (IOException e) {
             throw new RuntimeException("Failed to delete image file", e);
         }
 
-        // Delete the image entity from the database
+
         imageRepo.deleteById(id);
     }
 }
